move make_spend (todo: update references)
