+++ chia/types/coin_spend.py
@@ -1,39 +1,14 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Any, Union
+from typing import Any
 
 from chia_rs import CoinSpend
 
-from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.serialized_program import SerializedProgram
 from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
 from chia.util.streamable import Streamable, streamable
-from chia.wallet.program import Program
-
-
-def make_spend(
-    coin: Coin,
-    puzzle_reveal: Union[Program, SerializedProgram],
-    solution: Union[Program, SerializedProgram],
-) -> CoinSpend:
-    pr: SerializedProgram
-    sol: SerializedProgram
-    if isinstance(puzzle_reveal, SerializedProgram):
-        pr = puzzle_reveal
-    elif isinstance(puzzle_reveal, Program):
-        pr = puzzle_reveal.to_serialized()
-    else:
-        raise ValueError("Only [SerializedProgram, Program] supported for puzzle reveal")
-    if isinstance(solution, SerializedProgram):
-        sol = solution
-    elif isinstance(solution, Program):
-        sol = solution.to_serialized()
-    else:
-        raise ValueError("Only [SerializedProgram, Program] supported for solution")
-
-    return CoinSpend(coin, pr, sol)
 
 
 @streamable
